# docs/api/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "A location tracking API built with Go/PocketBase backend",
        "title": "Vibe Tracker API",
        "contact": {
            "name": "API Support",
            "email": "support@vibetracker.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8090",
    "basePath": "/api",
    "paths": {
...
        "/track": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "TokenAuth": []
                    }
                ],
                "description": "Tracks user location using GET request with query parameters",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tracking"
                ],
                "summary": "Track location (GET)",
                "parameters": [
                    {
                        "type": "number",
                        "format": "float64",
                        "description": "Latitude (-90 to 90)",
                        "name": "lat",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "format": "float64",
                        "description": "Longitude (-180 to 180)",
                        "name": "lon",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "format": "float64",
                        "description": "Altitude in meters",
                        "name": "alt",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "format": "float64",
                        "description": "Speed in m/s",
                        "name": "speed",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Unix timestamp",
                        "name": "timestamp",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Session name",
                        "name": "session",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Location tracked successfully",
                        "schema": {
                            "$ref": "#/definitions/models.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Authentication required",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "TokenAuth": []
                    }
                ],
                "description": "Tracks user location using POST request with JSON payload",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tracking"
                ],
                "summary": "Track location (POST)",
                "parameters": [
                    {
                        "description": "Location data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.LocationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Location tracked successfully",
                        "schema": {
                            "$ref": "#/definitions/models.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Authentication required",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "details": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "models.Geometry": {
            "type": "object",
            "required": [
                "coordinates",
                "type"
            ],
            "properties": {
                "coordinates": {
                    "type": "array",
                    "maxItems": 3,
                    "minItems": 2,
                    "items": {
                        "type": "number"
                    }
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "Point"
                    ]
                }
            }
        },
        "models.LocationProperties": {
            "type": "object",
            "required": [
                "timestamp"
            ],
            "properties": {
                "heart_rate": {
                    "type": "number",
                    "maximum": 300,
                    "minimum": 0
                },
                "session": {
                    "type": "string",
                    "maxLength": 100
                },
                "session_title": {
                    "type": "string"
                },
                "speed": {
                    "type": "number",
                    "minimum": 0
                },
                "timestamp": {
                    "type": "integer",
                    "minimum": 0
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "models.LocationRequest": {
            "type": "object",
            "required": [
                "geometry",
                "properties",
                "type"
            ],
            "properties": {
                "geometry": {
                    "$ref": "#/definitions/models.Geometry"
                },
                "properties": {
                    "$ref": "#/definitions/models.LocationProperties"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "Feature"
                    ]
                }
            }
        },
...
        "models.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "TokenAuth": {
            "description": "Custom token for location tracking endpoints",
            "type": "apiKey",
            "name": "token",
            "in": "query"
        }
    }
}